CDAP-4421: Provide Hive tokens to twill containers
