+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/runtime/distributed/DistributedMapReduceProgramRunner.java
@@ -26,6 +26,7 @@ import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.internal.app.runtime.ProgramOptionConstants;
 import co.cask.cdap.internal.app.runtime.batch.distributed.MapReduceContainerHelper;
 import co.cask.cdap.proto.ProgramType;
+import co.cask.cdap.security.TokenSecureStoreUpdater;
 import com.google.common.base.Preconditions;
 import com.google.inject.Inject;
 import org.apache.hadoop.yarn.conf.YarnConfiguration;
@@ -50,8 +51,9 @@ public final class DistributedMapReduceProgramRunner extends AbstractDistributed
 
   @Inject
   public DistributedMapReduceProgramRunner(TwillRunner twillRunner, LocationFactory locationFactory,
-                                           YarnConfiguration hConf, CConfiguration cConf) {
-    super(twillRunner, locationFactory, hConf, cConf);
+                                           YarnConfiguration hConf, CConfiguration cConf,
+                                           TokenSecureStoreUpdater tokenSecureStoreUpdater) {
+    super(twillRunner, locationFactory, hConf, cConf, tokenSecureStoreUpdater);
   }
 
   @Override
